# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Project Service API
  description: API documentation for ProjectService, including projects, vacancies, subprojects, stages, moments, and internships.
  version: 1.0.0
servers:
  - url: http://localhost:8082/api/v1
paths:
  /project/{projectId}:
    get:
      summary: Get project by ID
      operationId: getProject
      parameters:
        - name: projectId
          in: path
          required: true
          description: ID of the project
          schema:
            type: integer
      responses:
        '200':
          description: Project object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
...
  /subproject:
    post:
      summary: Create a subproject
      operationId: createSubProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubProjectDto'
      responses:
        '201':
          description: SubProject created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDto'
...
components:
  schemas:
    ProjectDto:
      type: object
      description: DTO for representing project details
      properties:
        id:
          type: integer
        name:
          type: string
          description: Name of the project
        description:
          type: string
          description: Detailed description of the project
        ownerId:
          type: integer
          description: ID of the project owner
        parentProjectId:
          type: integer
          description: ID of the parent project (if applicable)
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the project was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the project was last updated
        status:
          type: string
          description: Current status of the project
        visibility:
          type: string
          description: Visibility level of the project
...
    SubProjectDto:
      type: object
      description: DTO for creating or updating subprojects
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
          description: Name of the subproject
...